chore: remove MyService.DBus.cs
