--- src/Uno.UI.Runtime.Skia.X11/DBus/MyService.DBus.cs
@@ -1,132 +1,0 @@
-// Generated using `dotnet dbus codegen --protocol-api /usr/share/dbus-1/interfaces/org.freedesktop.portal.Request.xml`
-// Arch Linux's xdg-desktop-portal package v1.18.4-1
-// tmds.dbus.tool 0.21.2
-
-using System;
-using Tmds.DBus.Protocol;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace Uno.WinUI.Runtime.Skia.X11.DBus
-{
-	// From https://flatpak.github.io/xdg-desktop-portal/docs/doc-org.freedesktop.portal.Request.html#signals
-	enum Response : uint
-	{
-		Success = 0,
-		UserCancelled = 1,
-		Other = 2
-	}
-
-    partial class Request : MyServiceObject
-    {
-        private const string __Interface = "org.freedesktop.portal.Request";
-        public Request(MyServiceService service, ObjectPath path) : base(service, path)
-        { }
-        public Task CloseAsync()
-        {
-            return this.Connection.CallMethodAsync(CreateMessage());
-            MessageBuffer CreateMessage()
-            {
-                var writer = this.Connection.GetMessageWriter();
-                writer.WriteMethodCallHeader(
-                    destination: Service.Destination,
-                    path: Path,
-                    @interface: __Interface,
-                    member: "Close");
-                return writer.CreateMessage();
-            }
-        }
-        public ValueTask<IDisposable> WatchResponseAsync(Action<Exception?, (uint Response, Dictionary<string, VariantValue> Results)> handler, bool emitOnCapturedContext = true, ObserverFlags flags = ObserverFlags.None)
-            => base.WatchSignalAsync(Service.Destination, __Interface, Path, "Response", (Message m, object? s) => ReadMessage_uaesv(m, (MyServiceObject)s!), handler, emitOnCapturedContext, flags);
-    }
-    partial class MyServiceService
-    {
-        public Tmds.DBus.Protocol.Connection Connection { get; }
-        public string Destination { get; }
-        public MyServiceService(Tmds.DBus.Protocol.Connection connection, string destination)
-            => (Connection, Destination) = (connection, destination);
-        public Request CreateRequest(ObjectPath path) => new Request(this, path);
-    }
-    class MyServiceObject
-    {
-        public MyServiceService Service { get; }
-        public ObjectPath Path { get; }
-        protected Tmds.DBus.Protocol.Connection Connection => Service.Connection;
-        protected MyServiceObject(MyServiceService service, ObjectPath path)
-            => (Service, Path) = (service, path);
-        protected MessageBuffer CreateGetPropertyMessage(string @interface, string property)
-        {
-            var writer = this.Connection.GetMessageWriter();
-            writer.WriteMethodCallHeader(
-                destination: Service.Destination,
-                path: Path,
-                @interface: "org.freedesktop.DBus.Properties",
-                signature: "ss",
-                member: "Get");
-            writer.WriteString(@interface);
-            writer.WriteString(property);
-            return writer.CreateMessage();
-        }
-        protected MessageBuffer CreateGetAllPropertiesMessage(string @interface)
-        {
-            var writer = this.Connection.GetMessageWriter();
-            writer.WriteMethodCallHeader(
-                destination: Service.Destination,
-                path: Path,
-                @interface: "org.freedesktop.DBus.Properties",
-                signature: "s",
-                member: "GetAll");
-            writer.WriteString(@interface);
-            return writer.CreateMessage();
-        }
-        protected ValueTask<IDisposable> WatchPropertiesChangedAsync<TProperties>(string @interface, MessageValueReader<PropertyChanges<TProperties>> reader, Action<Exception?, PropertyChanges<TProperties>> handler, bool emitOnCapturedContext, ObserverFlags flags)
-        {
-            var rule = new MatchRule
-            {
-                Type = MessageType.Signal,
-                Sender = Service.Destination,
-                Path = Path,
-                Interface = "org.freedesktop.DBus.Properties",
-                Member = "PropertiesChanged",
-                Arg0 = @interface
-            };
-            return this.Connection.AddMatchAsync(rule, reader,
-                                                    (Exception? ex, PropertyChanges<TProperties> changes, object? rs, object? hs) => ((Action<Exception?, PropertyChanges<TProperties>>)hs!).Invoke(ex, changes),
-                                                    this, handler, emitOnCapturedContext, flags);
-        }
-        public ValueTask<IDisposable> WatchSignalAsync<TArg>(string sender, string @interface, ObjectPath path, string signal, MessageValueReader<TArg> reader, Action<Exception?, TArg> handler, bool emitOnCapturedContext, ObserverFlags flags)
-        {
-            var rule = new MatchRule
-            {
-                Type = MessageType.Signal,
-                Sender = sender,
-                Path = path,
-                Member = signal,
-                Interface = @interface
-            };
-            return this.Connection.AddMatchAsync(rule, reader,
-                                                    (Exception? ex, TArg arg, object? rs, object? hs) => ((Action<Exception?, TArg>)hs!).Invoke(ex, arg),
-                                                    this, handler, emitOnCapturedContext, flags);
-        }
-        public ValueTask<IDisposable> WatchSignalAsync(string sender, string @interface, ObjectPath path, string signal, Action<Exception?> handler, bool emitOnCapturedContext, ObserverFlags flags)
-        {
-            var rule = new MatchRule
-            {
-                Type = MessageType.Signal,
-                Sender = sender,
-                Path = path,
-                Member = signal,
-                Interface = @interface
-            };
-            return this.Connection.AddMatchAsync<object>(rule, (Message message, object? state) => null!,
-                                                            (Exception? ex, object v, object? rs, object? hs) => ((Action<Exception?>)hs!).Invoke(ex), this, handler, emitOnCapturedContext, flags);
-        }
-        protected static (uint, Dictionary<string, VariantValue>) ReadMessage_uaesv(Message message, MyServiceObject _)
-        {
-            var reader = message.GetBodyReader();
-            var arg0 = reader.ReadUInt32();
-            var arg1 = reader.ReadDictionaryOfStringToVariantValue();
-            return (arg0, arg1);
-        }
-    }
-}
